refactor: 메뉴입력 기능 리팩터링 메뉴 번호를 감싸는 MenuNumber 래퍼 클래스를 중첩 클래스에서 탑 레벨 클래스로 변경 InputView 에서만 사용되지 않기 때문에 탑 레벨 클래스로 분리 시키는 것이 맞다고 판단됨
